tests/wpt/web-platform-tests/css/css-backgrounds/parsing/border-image-outset-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Backgrounds and Borders: getComputedStyle().borderImageOutset</title>
<link rel="help" href="https://drafts.csswg.org/css-backgrounds/#border-image-outset">
<meta name="assert" content="border-image-outset computed value is four values, each a number or absolute length.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
<style>
  #target {
    font-size: 40px;
  }
</style>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("border-image-outset", "1px");
test_computed_value("border-image-outset", "1px 2");
test_computed_value("border-image-outset", "1px 2 3px");
test_computed_value("border-image-outset", "1px 2 3px 4");
test_computed_value("border-image-outset", "0 calc(0.5em + 10px) 3 calc(-0.5em + 10px)", "0 30px 3 0px");
</script>
</body>
</html>
